Change  name 'calendar; to 'event' all files,
