+++ lib/app/features/home/pages/calendar_page/add_event_dialog/add_event_dialog.dart
@@ -6,7 +6,7 @@ import 'package:primary_school/app/features/home/pages/calendar_page/add_event_d
 import 'package:primary_school/app/features/home/pages/calendar_page/add_event_dialog/cubit/add_event_cubit.dart';
 import 'package:primary_school/constans/colors.dart';
 import 'package:primary_school/domain/models/event_model/event_model.dart';
-import 'package:primary_school/domain/repositories/calendar/events_repository.dart';
+import 'package:primary_school/domain/repositories/event/events_repository.dart';
 
 class AddEventDialog extends StatefulWidget {
   const AddEventDialog({
