+++ lib/data/remote_data_sources/event_remote_data_source.dart
@@ -1,0 +1,3 @@
+class EventRemoteDataSource{
+  
+}
